pyproject.toml
[project]
name = "PgxClassicEnvs"
version = "0.1.0"
description = "Vectorized ConnectFour and TicTacToe examples using JAX"
authors = [{name = "Brandon Corfman", email = "h9tbgnbbk@privaterelay.appleid.com"}]
dependencies = [
    "jaxlib>=0.4.28",
    "pgx>=2.0.1",
    "mctx>=0.0.5",
    "solara>=1.33.0",
]
license = "MIT"
readme = "README.md"
requires-python = ">= 3.9"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.2.2",
    "pytest-cov>=5.0.0",
    "ipython>=8.18.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pgx-classicenvs"]

[tool.pytest.ini_options]
addopts = "--cov-branch --cov-report term --cov=src"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
markers=["unit: Unit tests.",
         "system: System (browser-based) tests."]
         
[tool.coverage.run]
source = ["src"]

